sensors/bmi270: add support for the new sensor framework
bmi270 can be used with the new sensor framework
