+++ drivers/sensors/Kconfig
@@ -223,6 +223,35 @@ config SENSORS_BMI270
 
 if SENSORS_BMI270
 
+config SENSORS_BMI270_UORB
+	bool "Bosch BMI270 UORB"
+	default n
+
+if SENSORS_BMI270_UORB
+
+config SENSORS_BMI270_POLL
+	bool "Enables polling sensor data"
+	default n
+	---help---
+		Enables polling of sensor.
+
+config SENSORS_BMI270_POLL_INTERVAL
+	int "Polling interval in microseconds, default 1 sec"
+	depends on SENSORS_BMI270_POLL
+	default 1000000
+	range 0 4294967295
+	---help---
+		The interval until a new sensor measurement will be triggered.
+
+config SENSORS_BMI270_THREAD_STACKSIZE
+	int "Worker thread stack size"
+	depends on SENSORS_BMI270_POLL
+	default 1024
+	---help---
+		The stack size for the worker thread.
+
+endif #SENSORS_BMI270_UORB
+
 choice
 	prompt "BMI270 Interface"
 	default SENSORS_BMI270_SPI
